Move disk and interface metrics to use 64bit OIDs
The disk partitions were only recording the oid for dfHighTotalKBytes, which only makes sense if the partitions are at least 2TB in size, otherwise they'll always report 0. Move them to the full 64bit OID, so that small partitions are also recorded.

Similarly move the network interface stats to use a 64bit OID, so that traffic on low usage interfaces appears.

Also, some tidying of the entries under sysStat so that they're declared as standalone metrics, and not ones in a table.