Move NetApp disk, interface metrics to 64bit OIDs (#25)
* Move disk and interface metrics to use 64bit OIDs

The disk partitions were only recording the oid for dfHighTotalKBytes, which only makes sense if the partitions are at least 2TB in size, otherwise they'll always report 0. Move them to the full 64bit OID, so that small partitions are also recorded.

Similarly move the network interface stats to use a 64bit OID, so that traffic on low usage interfaces appears.

Also, some tidying of the entries under sysStat so that they're declared as standalone metrics, and not ones in a table.

* Remove the provider: line

Removing the `provider:` line, as the matching entity definition in NewRelic is specific to NetGear, rather than a generic NAS box, so is a bad choice.

* Implement changes and fixes suggested by @i3149

Pull in the changes offered by @i3149 around losing the "net" prefix on ifDescr and ifIndex; updating the snapvault_ tags and making svState into a metric, not a tag.